requirements.txt
# PPT Master 依赖
# ================
#
# 大部分工具仅使用 Python 标准库，以下为可选依赖。
#
# 安装方式：
#   pip install -r requirements.txt

# ─────────────────────────────────────────────────────────────
# SVG 转 PPTX 工具 (tools/svg_to_pptx.py)
# ─────────────────────────────────────────────────────────────
# 用于将 SVG 文件批量转换为 PowerPoint 演示文稿
# SVG 以原生矢量格式嵌入，需要 PowerPoint 2016+ 查看
python-pptx>=0.6.21

# ─────────────────────────────────────────────────────────────
# 图片处理工具
# ─────────────────────────────────────────────────────────────
# fix_image_aspect.py - 修复图片宽高比（防止 PPT 转形状时拉伸）
# gemini_watermark_remover.py - 去除 Gemini 生成图片的水印
Pillow>=9.0.0
numpy>=1.20.0

